requirements.txt
Flask==2.3.2
python-docx==0.8.11
PyMySQL==1.0.2
Flask-SQLAlchemy==3.0.3
python-ldap==3.4.3
gunicorn==20.1.0
python-dotenv==1.0.0
reportlab==4.0.4
